Making dialogBoxes unclickable
